// src/app/models/spotify/track.ts
import {Album} from './album';
import {Artist} from './artist';
import {ExternalIds} from './external-ids';
import {ExternalUrls} from './external-urls';

export interface Track {
    album: Album;
    artists: Artist[];
    available_markets: string[];
    disc_number: number;
    duration_ms: number;
    explicit: boolean;
    external_ids: ExternalIds;
    external_urls: ExternalUrls;
    href: string;
    id: string;
    is_local: boolean;
    name: string;
    popularity: number;
    preview_url: string;
    track_number: number;
    type: string;
    uri: string;
}